all tests ok for now

+++ ft_printarg.c
@@ -6,7 +6,7 @@
 /*   By: glucken <glucken@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 16:27:19 by glucken           #+#    #+#             */
-/*   Updated: 2025/10/04 19:26:17 by glucken          ###   ########.fr       */
+/*   Updated: 2025/10/08 22:26:35 by glucken          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,14 +26,14 @@ int ft_printarg(char c, va_list *args)
 	if (c == 'i')
 		size+= ft_putnbr(va_arg(*args, int), 0);
 	if (c == 'p')
-		size+= ft_putptr(va_arg(*args, long int));
+		size+= ft_putptr(va_arg(*args, long long unsigned int));
 	if (c == 's')
 		size+= ft_putstr(va_arg(*args, char*));
 	if (c == 'u')
 		size+= ft_putnbr_u(va_arg(*args, unsigned int), 0);
 	if (c == 'x')
-		size+= ft_putnbr_hexa(va_arg(*args, unsigned long int), 1, 0);
+		size+= ft_putnbr_hexa(va_arg(*args, unsigned int), 1, 0);
 	if (c == 'X')
-		size+= ft_putnbr_hexa(va_arg(*args, unsigned long int), 0, 0);
+		size+= ft_putnbr_hexa(va_arg(*args, unsigned int), 0, 0);
 	return (size);
 }
